update domain carts
- Update Result
- Update domain models

+++ domain/domain-carts/src/main/java/com/study/domain/carts/models/Cart.kt
@@ -10,5 +10,5 @@ data class Cart(
     @ColumnInfo(name = "title") val title: String,
     @ColumnInfo(name = "description") val description: String,
     @ColumnInfo(name = "price") val price: Float,
-    @ColumnInfo(name = "amount") val amount: Float,
+    @ColumnInfo(name = "amount") val amount: Int,
 )
